make: rework how vendor-specific and BSPACM archives are managed
Assume that each vendor provides some sort of library to assist with
peripheral configuration.  Add the rules for that in Makefile.series,
and place the generated archive in the board directory.

Also add an archive for BSPACM sources, excluding startup files, that
need to be built once for each board and can be shared thereafter.

Eliminate use of VPATH.  Update the build rules so startup-related
material is placed in the current (application) directory, since it may
involve flags that are application-specific.