system_TM4C: copy from CMSIS 2.0
Except for conversion from DOS format line endings and renaming, this is
an unmodified copy of Device/ARM/ARMCM4/Source/system_ARMCM4.c from
CMSIS-SP-00300-r3p2-00rel1.zip.  It will serve as the basis of a CMSIS
system layer for TM4C since TI does not provide one.

(In particular, we need to configure the floating point unit on power-up
so that constructors that might reference it will not generate a hard
fault prior to main() being invoked.)