Makefile.common: fix dependency issues
A type inhibited generation of dependency files.  Also change so that
files contributing to libbspacm.a have dependencies generated, and that
the object files for that library are not removed in a standard clean
operation: that forces the library to be rebuilt, when we (I) normally
intend "clean" to mean removing only the application-specific material.