cleanup: add CC0 markings
Specifically for things that I expect somebody might want to
clone-and-own for experimental purposes, and that they may evolve to
something that has copyrightable IP in it.  No reason to confuse their
rights just because it derived from something I did that was trivial in
its original form.

Leaving this off some documentation for now, and off vector table and
memory scripts just because.

Find files that may not make their distribution restrictions clear by
using this command:

( git ls-tree -r --name-only HEAD ;
  git grep -E -l 'Copyright|publicdomain' ) \
| sort \
| uniq -u