core include: add device and config headers
The device header will be used to decouple bspacm/core.h from the set of
supported device series.  It also provides an isolated location to put
extension declarations, such as the upcoming typedef for the common
parts of GPIOA_TypeDef and GPIOA_AHB_TypeDef on TM4C.

The config.h header is intended to provide information about material
that is put in the periph_config.c file.  At this time it is merely a
placeholder.