cleanup: remove namespace prefix from source file names
We're about to change the build infrastructure so the originating
directory is preserved in the object file path, rather than dumping
everything into the board root directory where files with the same name
might provide different contents from different layers.  Consequently,
the namespace element of the base name of the file can be removed.

We're also introducing a new practice of using the underscore visibility
suffix to distinguish source files that come from the device hierarchy
rather than the generic device-agnostic hierarchy.