Makefile.common: refactor generated library linkage
In an upcoming commit we need to be able to better control searches
among application, bspacm, and newlib libraries.  Make this easier by
splitting the generated archive support into two pieces: one that is the
archive file (which is an order-dependency), and one which is the
library reference (which goes into LDLIBS).