config: rework to permit excluding UART support
Building nop was sucking in a bunch of code and data because the default
periph_config sources declared at least one UART per board.  Rework the
infrastructure so there's a common <bspacm/config.h> that sets defaults,
then includes a board/application-specific <bspacm/appconf.h> that
applies those defaults in a manner consistent with what's in the
board/application-specific periph_config.c.

Yes, it probably should be periph_config.h instead of appconf.h, except
that maybe sometime something other than peripherals will be part of
this.

While we're at it, get rid of the console abstraction for now, and just
use the default UART implicitly.