sbrk: rework how default is provided
nosys provided a non-functional default, while the mechanism for
specifying SBRK overrides the link-time garbage collection putting
_sbrk() in applications that don't use it.  The new approach restores
"optimal" behavior: if _sbrk() is unreferenced nothing provides it, but
if it is referenced a working implementation will be provided by
default.

This gets the size of nop back to where it was a week ago before we
started getting tricky.