core: add support for bitband SRAM and PERIPH access
Only supported on M3/M4 devices, not M0+.  Not useful for setting
multiple bits simultaneously.  Useful for manipulating individual
peripheral register bits when the bit and the register address are
compile-time constants, and when setting flags in an event record while
interrupts are enabled.

Note that peripheral registers are always accessed as 32-bit objects,
while SRAM objects may be accessed as smaller entities.