fdops: remove fcntl(2) support
The newlib build flags for ARM exclude HAVE_FCNTL, so newlib provides a
strong definition of this function that returns -1/ENOSYS, overriding
the definition here.  Since fcntl() operations are the purview of libc,
we'll remove the override.  It becomes more likely that ioctl will be
needed, then.

This also resolves the question of whether O_NONBLOCK, O_ASYNC, or
O_DIRECT might be usable: no.