uart: rework transmission architecture
In support of low-energy use we want the application to have more
control over what to do when resources are unavailable.

The UART write operation will no longer ever block or enable interrupts
in order to transmit something.  The bridge to newlib will do this if
necessary to satisfy newlib's requirements.  Interrupts are now disabled
during transmit only to protect tx_fifo_ni_ and tx_state_.  UART FIFOs
are now optional, eliminating space overhead if programmed IO or DMA are
intended to be used.  FIFOs if present are accessed only when interrupts
are disabled.