efm32 uart: refactor configuration
The previous solution separated the UART pinmux config from the material
that could be inferred from the peripheral name and instance.  The
pinmuxUART structure didn't really belong in the gpio header, and having
two configuration structures was complex, as was the need to scan for
the correct pinmux config for a given instance.

Replace this with a single and complete configuration structure per
peripheral.  Factor out the common parts shared by all UARTs, and make a
stronger distinction among UART, USART, and LEUART (which will simplify
upcoming SPI support).

The BSPACM UART structure still references this device-specific
configuration, but the default is now a weak reference, theoretically
allowing applications to change the pin configuration without having to
clone periph_cfg.c.  We'll find out if this was a good idea.

This commit addresses the issues for the EFM32GG-STK3700 only.  After
some additional refactoring the remaining boards will be updated.