tm4c: enhance GPIO interface
Provide inline functions to convert from GPIO peripheral addresses to
the corresponding bit in registers like RCGCGPIO; similarly for
diagnostics to get the letter associated with the instance.

Where the port is a compile-time constant the shift value can be
resolved at point-of-call, allowing very efficient code through
bitbanding the configuration registers.