tm4c/gpio: eliminate .port_shift field from pinmux
Forcing the user to provide this creates too great a risk of
inconsistent configuration.  Suck it up and accept the cost of
dynamically determining the shift from the port at runtime, in those
cases where the inline function that does this isn't given a
compile-time constant to use in the comparisons.