nrf51: stub for soft-device support
Existing Make NRF_SOFT_DEVICE specifies the soft device used in linking.

Add NRF_USE_SD to indicate that any present soft device is intended to
be used.  The expectation is that any non-SD-aware application can link
with the soft-device and run without engaging it.  By default if there's
an SD linked in it's assumed to be used; override with NRF_USE_SD=0.

Add NRF51_SD_HEADER_PATH analogous to NRF_SDK_ROOT to tell the build
infrastructure where to find soft-device headers.  This is used only
when NRF_USE_SD is non-zero.

Add functions that invoke the soft-device interface wrappers instead of
the standard CMSIS functions for operations that the soft device wants
to be aware of.  Selection of implementation is determined by
NRF_USE_SD.  Only NVIC-related functions are defined in this commit.

NOTE: The S110 v7.1.0 softdevice faults if NVIC-related functions are
used with peripherals not restricted by the softdevice.