nrf51: implement delegation to soft-device wrappers
This gets ugly because the wrappers only function when the soft-device
is enabled and the requested peripheral is restricted by the SD.  This
means we need to check for a disabled SD and fall back to the non-SD
implementation in that case.  Coupled with the need to provide the
non-SD implementation when SD support is not enabled at build time this
makes for ugly code.

FWIW: Enabling the soft-device appears to take about 1ms and disabling
it takes about 250us.  The latter time might increase if the device
was actually configured to do something.